// src/client/app/_models/studyAreas.ts
import { Injectable } from '@angular/core';

@Injectable()
export class StudyAreas {

   idListaAreaEstudio: number;
   codigoListaAreaEstudio: string;
   nombreListaAreaEstudio: string;
   ordenListaAreaEstudio: number;
   indicadorHabilitado: boolean;
   auditoriaUsuario: number;
   auditoriaFecha: string;
   label: string;
   value: number;

}